Corrected the errors to make the intensity correction work.
Now, when stitching, the user is able to decide whether they want the images to be corrected or not simply by putting the argument  *correction* of the function *stitching_scrapbooking_allImages()* in stitching.py to True (meaning that the Gaussian enveloppe of the intensity in the images is corrected). If correction == False, the images are not intensity-corrected.