Trying to modify the images for vertical shift estimation with noise, but doesn't work .
I thought it would work to make the non-overlapping regions in black and add some random noise to it to prevent the phase cross-correlation function to match those regions and estimate the shift of the black regions instead of the structures of the image, but it doesn't work at all, sadly. It is still estimating the shift of the black region.
I'll try later with cropping and adding noise to only one image instead of both.

Also, added a subtract function in the imageTreatment to subtract a value to all pixels.