Testing the code for verification of its stitching
Seems that the code calculates the shift with the stage position in the file name very well, but it is the way the files are named that confuses the stitching. The code assumes that the files are numbered in order, which is true, but the code also considers that 10 000 is smaller than 1 000 because it has more zeros. lol
I had to rename the files to make it work, but this way is definitely something I need to change eventually.